+++ src/main/kotlin/mine/sweeper/SweeperMap.kt
@@ -1,16 +1,16 @@
 package mine.sweeper
 
-class SweeperMap(private val option: MapOption) {
-    private val fields: Fields = Fields(option)
-    private val randomPosition: MutableList<Pair<Int, Int>> = option.randomPosition()
+class SweeperMap(setup: SweeperMapSetup) {
+    private val fields: Fields = Fields(setup.initFields())
+    private val randomPositions: MutableList<Pair<Int, Int>> = setup.initRandomPositions()
 
     fun entireMap(): Array<Array<Field>> {
         return fields.entire()
     }
 
     fun randomField(): Pair<Int, Int> {
-        check(randomPosition.isNotEmpty())
-        return randomPosition.removeFirst()
+        check(randomPositions.isNotEmpty())
+        return randomPositions.removeFirst()
     }
 
     fun setMine(height: Int, width: Int) {
